refactor(single requests): Refactor results filter
This refactors the way the filters are built and sent to the view layer.

Previously this was using `res.locals` but in an attempt to move away
from that method this now sets values on the request object and allows
the rendering controller to pass things down to the view.