refactor(single requests): Change how single requests are set
This refactor changes how "moves" are set. Currently they are stored
on `res.locals` and passed to the view that way, but this starts to
adopt the pattern of avoiding settings things on `res.locals` in lots
of middleware and using them in following middleware.

Splitting out how the results are set will also allow the templates to
be more shareable as they expect more generic values to exist on the
request object that they can then in turn pass down to the view.

It also helps to address #451.