build: Run end-to-end tests on pull requests
This allows us to run the end-to-end tests against Heroku Review Apps
on new Pull Requests.

The con of this is it may increase the time tests take to run on
PRs.

However, the pro is that we will know sooner than merges to master if
our changes are breaking any existing journeys that are tested in the
end-to-end tests.