fix: Only display move agreed status for single requests
The logic for displaying this label and value in the summary panel
was checking if it was not undefined. However, the API currently returns
a value of `false` and does not support `null`.

This meant that for all existing court moves this label was being displayed
as "Not agreed". This information is not relevant for these types of moves
so the logic has been updated to only display for prison transfers.