refactor: Use common body for filters and results
This refactor moves the shared body of filtering a set of results
and displaying the filter to a common middleware that preceeds both
middleware responsible for setting the filter and the results.

This means that we reduce the places we need to have this code and
can move in a direction to share the setting of filters and results
going forward.

This change also amends how single requests are "filtered". Previously
this was done as part of the URL path which meant a bit more heavy
lifting around middleware. This moves to a query based approach
which means we can share the code across other parts of the application
to handle filtering and sorting.