refactor: Stop using res.locals for date range
This is part of a wider piece of work to stop using `res.locals`
through middleware. This technique makes it harder to see where something
in a view is coming from.

This also refactors some of the date utils to make use of app wide
configuration and improve how dates are handled.