chore: Support default query handling of collections
This adds support to define a default query for a list of things, or
"a collection". In this case, a list of moves.

We will need to start supporting more things in the query, such as
sorting and possibly other filters. This starts to standardise how
we can return a list of things from the API based on certain criteria.