refactor(permissions): Update how users see all locations
This change refactors how we handle whether a user has permission to
see all locations or not. Previously we associated it with moves
because that was the only thing we listed.

Now we need to support a user to be able to sign in and see results
for all locations.

This refactor will allow us to use this permission to support a
Population Management User being able to see requested moves for
all locations.