test: Update end-to-end tests to use meta content for location
Previously the tests expected that when a user signed in, they were
taken to a URL that contain the location that was selected.

Now that we are introducing a dashboard where the user will land at
the root of the domain this method will no longer be possible.

This update adds a meta tag to the head containing the user's current
location ID. This allows it to be used be the end-to-end test suite
as well as anything that we may need to do in the client.