test: Reduce the speed of the editability tests
This tries to reduce the need to create a new move for each editability
test in order to speed them up.

It also moves away from moves being created based on a user and their role
as this is a concept we have tried to stay away from. We can keep some
checks linked to roles based around permissions but when creating those
fixtures they shouldn't be tied to a person's role.