refactor: Use unique key for request body in collections
Currently we set the request body as a single object but the structure
of that needs to be different for different result types.

This creates a namespace so that it doesn't always overwrite that value
if we need to use multiple middlewares for particular routes, for example
on a summary dashboard.