fix(moves): Cater for moves without a person in CSV download
Previously all moves were associated with a person. Now with allocation
moves, they can be created without a person which meant that when
attempting to download moves for a day that contained allocations without
an associated person it caused an error.

This fix ensures the assessment answers falls back to an array so that
the CSV provides a default value instead of causing an error.