fix(single requests): Remove allocations from single requests results
Single requests are being filtered on a `move_type` of `prison_transfer`
which was enough to distinguish them previously. However, we have added
support for allocation moves which are also classed as a `prison_transfer`.

This means we needed a further way to remove results related to an
allocation move from single requests.

This change adds the new filter within the API to exclude allocation
moves.