fix(allocations): Ensure person is formatted correctly in allocations
This uses the same method for formatting a person when it comes back
from the API. This add expected properties for particular parts of the
application like `fullname` and `image`.

It would be good to this every time that model is deserialized from the
API using devour but that would involve a large refactor at this stage
so this change re-uses a method already used within the moves service.