fix(components): Fix image size in card component
This fix ensures that the image doesn't change size depending on the
size of the content area. It was previously using flex layout which meant
that in cases where a person had a lot of tags the image on the left
would shrink and leave an uneven layout when there were multiple
cards displayed on a page.

This change uses a more conventional float based layout to ensure the
image always remains the same width.