refactor: Improve how a move without a person is handled
Previously if a move didn't contain a person we would return an empty
object. This caused some issues with particular logic where it wouldn't
treat that as an empty result.

This refactor ensures we returned undefined for a person object if there
isn't one on a move and we update the locations within the app where
we deal with a person and expected there to be one.

It also includes some updates to defaults when we don't know who a person
is in particular parts of the app like displaying them as a card.