fix(moves): Show placeholder card for moves without a person
Previously the display of a "card" component relied on a move having
a person and where mainly based around displaying information from that
person.

This fix ensures that there are certain fallbacks when a move doesn't
have a person and that we display a different style to help distinguish
that move in listings.