refactor(move): Move assinging of a person from allocation to move
This moves the position of the assign form wizard from being nested
under an `allocation` to being under a `move`. This action is happening
to a move, even though it's initial starting point is the allocation, so
it feels cleaner to have the URL be part of the move.

This means that any move can have the ability to assign a person to it.

It also makes it easier to get some of the information needed as most of
the data comes from the individual move anyway.

The other benefit is that we can provide an prompt on any move without
a person to assign somebody to it without needing to go to the allocation.
With the current implementation we would require them to go to the allocation
to click "add" where we would pick another random move instead of the
move they were just on.