fix: Ensure card without image doesn't have left margin
A previous update to the card component fixed the float issue with
images. This meant introducing a fixed width margin on the left.

However, this caused a problem when a card didn't include an image
as the margin was still being applied.

This fix uses a modifier class on the card for when an image is present
to handle applying the extra style required for the image.