refactor: Remove setting of locals from middleware
This helps to address the issue we have of setting values for the view
layer on `res.locals` within middleware.

This approach makes it a little harder to track where values within the
template are actually coming from so we have decided to take an approach
to define all the view variables in the controller that renders that
template.