refactor: Set default time periods within body middleware
This refactor moves the setting/overriding of the default time period
to the relevant body middlware.

The body is used to populate the filters on the dashboard and previously
we were overriding the date ranges in a middleware just on the dashboard.

When looking to add more panels to the dashboard this started to feel
repetitive and required a different date range.

So this refactor moves that responsibility to the relevant middleware
to handle instead. Then when the filter and body are included on the
dashboard route the default time period is used to display the results.