refactor: Improve location ID override for dashboard
This refactor improves the way the location ID is set for dashboard
filters.

Previously it updated the key in the body for each filter.

Now it sets `req.params.locationId` to the current location in the
session if it is set.

This means it can be more consistently shared between other middleware
that set the body based on the request params like in the collection
routes.