refactor: Fix naming of allocation presenters
The presenters used to turn an allocation into particular components
are currently named the wrong way around.

`allocationToMetaListComponent` actually builds the structure for a
`summaryList` component and calls the `govukSummaryList` component with
the result.

`allocationToSummaryListComponent` actually builds the structure for a
`metaList` component and calls the `govukMetaList` component with
the result.

This change switches the naming so that it matches correctly and also
updates the naming of the variables in the controller.