refactor: Update allocation view to use shared card component
This change updates the allocation template and controller so that
the card component can be more widely shared.

Currently it was doing a lot of manual lifting to create the view
for when a move doesn't have a person associated yet (a new allocation
move).

This change refactors that so it will more consistently inherit the
styles applied to the card component to create a bit more unity across
the application.