refactor: Update allocations and moves to use common date filter
This moves the logic for display a date along with the relative date,
if it matches, in brackets to the shared nunjucks filters so that
it can be used for both allocations and moves.

Currently the logic was in a move specific presenter which meant
the display of date was handled differently between the two views.

This moves that logic so that it can be more easily re-used.