fix: Ensure end-to-end test assertions are called with a promise
This method was assigning the result of a promise to a variable and not
using `await` to wait for the result. This fix ensures that the promise
will be waited for by passing it into TestCafe's `assert` method directly.

It also includes a small tweak to the variable names to make them a little
more readable.