fix: Remove `move_agreed` filter from allocation assign
This fix removes the checks from when to display the `move_agreed`
and `move_agreed_by` fields for a move.

Previously the API didn't support `null` values for these fields
which meant that on creation of new moves, it defaulted the value
to false. However, `false` actually means that the person has
answered that the move was not agreed with the receiving prison.

This broke in the case of moves within an allocation because this
question had never been asked as part of creating the move at any
point.

This supports the change in the API that supports `null` for this
field and updates the display based on that new logic.