fix: Update button text for assigning a person
This updates the logic that defines the button text on the final
step of assigning a person to a move to be more accurate to what is
happening.

Before this set of changes the text said "Continue", which is the
text used to suggest there are more steps to come.

This uses the existing button action to show that the final step will
then add a person to the move.