fix: Remove case transformation on locations page
Previously we were trying to change the case of the list of locations
on the locations page.

The problem with this is that it then doesn't replicate how we replay
the location everywhere else on the service, for example, the banner in
the header or the list of locations in dropdowns.

This change removes that fix so we at least replay them consistently.

If we want to amend the case, this should ideally be done at source
(the API or NOMIS), or in some form of translations for specific
locations.