test: Fix autocomplete helper to contain parent
This change uses a selector to find the correct parent for the
autocomplete element before selecting the options.

It produced an error where if provided with no value the helper to
select options would count all options on the page rather than just
for the selector.

That lead to cases where the random value was then outside of the options
and would cause failures.