chore: Increase end-to-end test timeout
This increase the default timeout for the TestCafe suite.

In some instances, like waiting for the next page after clicking a
link TestCafe fails an assertion as page loads can take a little while.

This increase that timeout to see if it will help some of the inconsistent
test failures we have seen recently.