feat(allocations): Remove single move from allocation
fix(allocations): Remove link from card with unallocated slot

Previously we showed the link to add a person on unallocated slots. But in the PMU view, you are not allowed to assign a person. This PR introduces the possibility of not displaying such link

Revert "fix(allocations): Remove link from card with unallocated slot"

This reverts commit d91a77ac694c2b0fab34ed91d27a2ffe22ed3a4a.