chore: Support depedent fields within the form wizard
This addition means that the form wizard will always set the correct
`dependent` property on a field if it is added as a conditional
field within another field's options.

Previously creating conditional fields was a two step process that
involved adding a field as a conditional option to a parent as well
as adding a depedent property to the child field to ensure validation
is only applied in the correct circumstances.

This change doesn't affect any previous implementations but makes this
process easier moving forward so that when dynamically creating fields,
like will be the case for the Person Escort Record framework, you
don't need to loop over the fields again to set the dependent values.
This is now taken care of within the form wizard itself.