refactor: Expose move on request object
Ideally we would like to move away from setting variables on
`res.locals` within middleware. A better pattern is to assign the
variable on the request object so that it can be used elsewhere within
the request cycle.

This starts to move away from that by making a move available on
`req` so that it can be used in another part of the application.

Continues to address #451