refactor: Add markdown support to before and after fields to form wizard
This moves the support to include markdown before or after the fields
on a form wizard step to the main form wizard template instead of just
allowing it within the framework.

This means that if a form wizard step inherits from the base template
it can still support display of markdown before or after the fields.