refactor: Use req instead of res.locals in move detail controller
This starts to move away from the issue described in #451 where we
have been setting values to `res.locals` in middleware and then
using them lots in the views.

It's cleaner and easier to trace if things are set from a controller
so this starts to move towards `req` over `res.locals`.

Continues to address #451