test: Check step URLs when creating an allocation
If there is an error on a step during the allocation process its
hard to immediately debug that the app didn't get to the correct step.

This adds in a check to explicity check the expected URL and more clearly
describe the error if it fails.

This was highlighted by an issue in the allocations tests where they
set the date to the Friday for that week. This makes the tests a little
brittle as they will only pass on 5 of the 7 days of the week. A fix
will be applied in a later commit to address this.