fix(e2e): Prevent allocation tests only running Monday to Friday
Currently the allocation tests use the Friday of the week as the
allocation date. This currently means if they're run on any other day
they'll fail.

It also currently causes us to have a large number of allocations for
the current week if the automated tests are run often and compounds
the current performance issues to display allocations.

This changes the allocation to use a date in the future and not for
the cancel fixtures to not require visiting the allocations dashboard
which has also caused timeout failures.