fix(e2e): Fix allocation validation checks
TestCafe had started to give a warning that particular tests were passing
an entire selector instead of the `.exists` property to check if an
element exists on the page.

This test wasn't working as expected so this change fixes it to use
the correct selector as well as DRYing up the implementation and
sticking to the Page Object approach of defining elements within
a Page.