build: Improve e2e parallelism
Rather than creating a separate testcafe instance for each and every test,
this splits the tests into the number of buckets matching the max_processes
and spins up that number of Testcafes. NB. if max_processes is set to 1,
then the behaviour is exactly as if calling the original non-parallelised
script.

Additionaly this has the side effect of making killing the process kill
all the child processes.

NB. If the process is killed before the browsers have connected to the
Testcafe runner server, those browsers will show as "Disconnected" and
have to be killed manually.