chore(per): Support ability to nest app
This adds the ability for the Person Escort Record application to be
nested if the request property has been defined within the request
cycle.

This gives us the ability to nest this app within a move as well as
allowing users to visit root URLs. For example:

`/move/:moveId/person-escort-record`
`/move/:moveId/person-escort-record`
`/move/:moveId/person-escort-record/:section`

The above is the equivalent to:

`person-escort-record/:recordId`
`person-escort-record/:recordId/:section`