chore(per): Update start/progress banners for PER on move detail
This change updates how the banners are displayed during different
states of the Person Escort Record.

Before a PER has been created the banner uses the instructional style
to make it clear an action needs to be taken.

Once it has been stared, it displays the progress of the different
sections of the PER.