fix(per): Remove before you start page when creating a new PER
This removes the conditional logic we had for before you start pages
being dependent on location type and needing to occur before creating
a new Person Escort Record.

The requirement is to have them come in before each section of
the journey so these have been removed and the creation happens
immediately. After creation, the user is redirect back to the move
detail page.