fix: Correct the contrast of links on light blue banner
This uses the shade mixin to inject a bit of black to the link colour
when links appear on the instructional banner that has the light blue
background.

The new colour is only marginally different but now passed with WCAG AA
on normal size text with a contrast of 4.55:1, from a previous value
of 3.99:1.