chore: Create setLocations middleware
Locations have been checked in various different places in an inconsistent fashion.

This aims to replace those duplicated checks in a consolidated middleware.

It checks the following properties in turn

1. req.params.locationId
2. req.session.currentLocation
3. req.session.currentRegion.locations
4. req.session.user.locations

and returns an array of ids

It is invoked after we have ensured that the user has selected a location of some sort