refactor: Use package scoping for framework
This updates the way the Book a secure move framewors repo is installed
to use NPM scoping.

This means that we can support multiple versions moving forward all
scoped within the same dependency grouping. Multi-version support will
be required once we start publishing new versions and existing versions
are not fully completed or need to be viewed.

This changes the build steps to accommodate the new folder structure
that is created in `node_modules/`.