refactor: Set move on req instead of res.locals
This helps to reduce the issue we have of things being set to `res.locals`
within middleware and making it a little harder to find where is it being
set.

This moves towards a pattern of setting things on the request object
within middleware if that value needs to be used later on in the request
chain.

Then if that value is required by a view, it is handled within the
controller that calls that template.

This change address removing `res.locals.allocation`.

Addresses #451