fix(frameworks): Ensure filter dependent responses correctly
This change ensures that any dependent responses are not sent to be
saved if the dependent field doesn't match the condition the field
depends on.

This also includes a fix to remove any falsy values from Array like
objects which prevents the API from returning an error when an
array with one empty string is sent when no checkboxes are selected.