refactor(per): Remove need to call PER service twice
This refactors the middleware to remove the unnecessary double API call
to get a Person Escort Record.

Previously the necessary `include` options weren't available on the
`/moves` endpoint so the frontend needed to make another call to get
the Person Escort Record.

This change removes that need by making use of the PER that is
associated with a move.