refactor: Extract assessment answer presentation
This change extracts the handling of rendering assessment answers
within another component, for example, a list of panels with a tag
into another presenter.

This will allow it to be shared by other parts of the app and give the
ability to present the assessment answers in separate ways more easily.

This change does still contain a style of loading settings internally
that we would like to get rid of (see #696) as the refactor would
be a much bigger and more time consuming change at this stage.