refactor: Display court information same way as risk information
This refactors the presenter to display assessment answers as a
summary list component to be the same format as the other presenters
and use an assessment category as the input.

This allows the two presenters to be swapped more easily depending
on the categories available. This change supports the work that is
needed to group the assessment answers differently once a Person Escort
Record has been started.