feat(per): Enable permissions for Person Escort Record
This creates the necessary permissions needed to access various
journeys within a Person Escort Record.

It uses the existing permissions model approach to assign permissions
based on roles from the Auth Service.

It wraps some pieces of the UI in logic and uses the protectRoute
middleware to prevent access to particular routes if the user
does not have access to those permissions.