chore(per): Add skeleton controller and view to print record
This creates the controller and view needed to view the Person Escort
Record in a print format.

It re-uses as many presenters from the view state of a move to reduce
code duplication.

For now this doesn't include a summary for the actual move or any
custom elements that are needed for the view state of the record.