fix: Ensure assessment answers are rendered correctly
A recent changed updated how assessment answers are transformed
in the presenter to load the child component.

This meant updating the way the components were called in the view and
in this instance it was missed.

This fix updates the way the component is called to display the
information correctly.