chore(per): Popup print dialog on page load for Person Escort Record
This adds a small amount of javascript to trigger the browser
print dialog on load of the print view of the Person Escort Record.

When users visit this route the only intention is to print this document
so by shortcutting them to the print dialog it makes this journey
easier.