chore: Add link back to move on Person Escort Record overview page
Add a link to the overview page of a framework section so that a user
can get back to the move after viewing a section.

Currently the user would hit a bit of a dead-end when viewing the overview
of a framework section, this allows them to navigate back to the move
detail page.