refactor: Update card component so PER tags can be switched with option
The logic to display the tags from the Pesron Escort Record or not
now needs to be based on a user's permissions. This means getting
access to the request object as well as environment variables which
requries a refactor in the way this presenter is called.

This change refactors the card component to support the condition to
be passed into the functions in a more pure way so that the logic can
be extracted a bit more.