chore(per): Fix page titles for Person Escort Record
No page titles were being set for the Person Escort Record pages
which meant that the browser tab and history would always show the same
title.

This isn't very useful when trying to navigate using the browsers
back or forward buttons or history so this change makes it clearer
what page a user was one.

It also helps with collecting page titles in analytics like GA.