fix(per): Ensure steps can't be accessed once reocrd is confirmed
This change prevents access to the framework steps once a Person Escort
Record has been confirmed.

The API already won't allow this so this just handles that journey
better on the frontend so that the user won't end up seeing API
error messages.

It also hides the actions from the overview if is has been completed.