fix(per): Ensure move detail logic uses permissions too
The logic for switching the display of the move detail page was
only being done on whether a Person Escort Record existed.

This was because this logic was introduced separately to the fix
to ensure user's without the correct permission don't see this feature.

This amends that condition to use the new controller variable to ensure
consistent display.