build: Preload key requests
This adds the key font files to the `head` as preload assets.

Preloading assets can improve page performance as the assets that are
known to be required can be asked for sooner in the network of requests.

We use these key fonts on all pages so asking for them before needing
to wait for the CSS to request them speeds up the page load times.