chore: Customise what includes we need for API requests
Currently we use a default include, specified on the model, that
asks for **every** include when making requests to the API for a
particular model.

This has started to make performance quite slow as certain requests,
like loading a list of moves or allocations, don't require the same
detail as when viewing a single resouce.

This change tailors the current includes used to ask for a list of
resources to only require the minimum to build that view.