refactor: Extract download middleware to separate function
We would like to tailor the include calls more to only what is needed
to build particular views or downloads.

Previously the app used the same middleware for both rendering and
download moves. However, this has caused a number of performance issues
so we now need to use separate calls.

THis also means needing to split the download middleware separately
so that it calls a different method in the service.