fix(framework): Don't use session values in form steps
Previously the framework step controller was extending the form wizard
session values.

However, for the this framework we always load in the values from the
API, unless there are errors.

This fix changes the `getValues` method to either return the error values
for when the page contains errors, or load the values only from the API.

This prevents a scenario where a conditional step has values, which
are then cleared out in the API. However, when a user navigates back
in the same session, that step contains values from the session which
override the values from the API.