fix(per): Group responses correctly under flag summary
Previously if the response type was `collection` and multiple values
were selected which correlate to a "tag" needing to be shown the
question text was being duplicated.

This was because the response values were being filtered and then
appending separately.

This fix checks to see if the response already exists when it's a
collection and appends to that response instead of creating a duplicate.