feat(frameworks): Load framework on runtime so support multiple verisons
The framework was previously loaded when the app started which meant
we were restricted to only using that one framework for all responses
from the API.

This was initially built this way because it was more performant for the
app to not load the framework on every request. However, to fully support
multiple versions we need to be able to support multiple frameworks
at one time and load the relevant questions/structure for the version
that is returned from the API.

This refactors the router to set the framework based on the version that
the API returns for that resource instead of always using the same
version.