chore: Add context for form wizard field options
If a form field has a context property, this will be passed as context
by `translateField` to i18n method.

Adds `setFieldContext` middleware to form wizard controller.

This sets the context of fields of the current step to that step's key.

NB. it is, of course, also possible to manually set a field's context

```
req.form.options.fields.foo.context = 'bar'
```

These changes enable the reuse of fields of the same name and prevents the creation
of indirect fields just to have a custom label or hint