chore(frameworks): Cache frameworks instead of always loading from disk
This adds a local cache for the frameworks service using the framework
name and version number to prevent needing to lookup the framework
from files every time a framework is asked for.

The framework is asked for frequently when interacting with the
Person Escort Record feature so this should prevent us having to access
the disk frequently when the app is running.