refactor: Set optional field label dynamically
This adds another field helper to determine whether a field is optional
or not and set the label suffix (`(optional)`) dynamically.

This change started in the framework but I moved it to the form wizard
to reduce the complexity of the importer and to share the logic across
all areas where we use the form wizard.