refactor(framework): Set initial values to session instead of to values
This moves the process of setting the initial values from the API
to the setup section of the lifecycle instead of within the `getValues`
method.

This allows the add another functionality to be supported alongside
this behaviour as it requires the session values to take precedent.

With the previous method there were too many crossovers causing incorrect
values to appear in particular scenarios.