feat(formwizard): Add support for "Add another" style fields
This adds support to the formwizard controller to handle a field type
that supports nested fields that can be repeated in a group.

It uses a similar pattern previously used to render conditional fields
to make it more dynamic and adds lifecycle events to the form wizard
to support the addition/removal of items.