chore: Use no locations for supplier if no current location or region
This ensures that location info is only sent when we want to filter
the moves for a supplier which are now correctly retrieved using
the supplier_id filter.

Previously we returned all the moves for the locations requested which
could have included moves that did belong to the supplier.

Furthermore, the list of supplier locations is now determined by where
the suuplier actually has moves and this can change dynamically, so
sending the locations meant that moves to new locations could be missed.