refactor: Update value types to use new API response values
The value types coming from the API have been updated to support a new
pattern that was introduced to cater for multiple types of collections
or objects.

This means that we need to update the references to them in the frontend
so that things still work correctly.

This change extends all calls to particular types to include both
the existing value plus the new value so that it can be released without
causing any breaking changes.