fix(framework): Add support for remaining nested response types
In order to meet the tight deadlines this reducer was only built to
cater for the nested response types that were needed for that release.

The code wasn't unit tested fully and didn't cater for the other
known response types at the time.

This improvement uses a circular approach to reducing the responses
from a nested response type, like `collection::add_multiple_items`,
in order to re-use as much as possible.

This change also includes unit test coverage to the missing lines.