refactor(framework): Always use type from question insted of response
With the addition of nested responses the client application had to
start looking in two different places to determine how to format a
response for the API.

This refactors the reducer to only use a single place - the question.

`response_type` was added to the question in order to build nested
responses as part of `collection::add_multiple_items` types.

This change tries to simplify the code that handles changes form
values into an API payload.