fix(single-requests): Make single request download work again
Single request downloads were still fetching the download data from the BE as JSON but then
attempted to send that data as if it were already in csv format but without processing it,
resulting in a circularity error.

This fix gets the download data from the BE endpoint as csv.