test: Update test prison number to be correct format
The downstream Prison API recently merged a change that expected the
search format to be in a pariticular format. If it isn't, it will now
throw an error which has been bubbled up from the API.

This means the frontend now receives a 500 from the API and doesn't
display the expected empty results page which causes the e2e tests to fail.

In order to unblock this for now, the unknown prison number has been
updated to be in the expected format.