fix(per): Prevent error when record already exists
An issue was recorded in sentry which returned an API error when the
frontend was attempting to create a new Person Escort Record when
one had already been created for this move.

This scenario could occur when:
- two users open a move without a record and see the start button
- first user clicks to start a new PER
- second user clicks to start a new PER but will receive an error
because one had already been created

This change will look for that error and then handle it more appropriately
by redirecting back to the move.