fix: Inherit from correct template on special vehicle message step
The template used to display the special vehicle message was inheriting
from a template that didn't contain the block being set in the template.

This led to the message not being displayed correctly.

This change updates the template inheritance correctly so that the
message will appear. It also fixes a stray `s` which could have come
from an attempted save shortcut.