fix(framework): Handle empty multiple items response types
This change adds a fix to display the "Answer question" text when
a response with a type of `add_multiple_items` has no items.

It will now fallback to the existing part of the presenter instead
of trying to do the custom logic for handling this response type.

This change also includes the addition of some missing unit tests.