fix(framework): Don't remove emdash or endash from values
A formatter which would replace emdashes and endashes with a hyphen,
which is gramatically incorrect in some instances, was causing
some questions to fail validation as the formatted value didn't
exist within the options for that field.

This removes that formatter from the list of defaults for the wizard
used for the framework so that the validation will correctly pass.