fix(per): Ensure nested fields are populated with initial values
The works to support add another fields, that can also be include
nested fields, meant re-working the order of some middleware.

This ended up causing a side-effect where nested fields weren't being
repopulated when editing a step.

This was because of the ordering of the middleware. The synthetic nested
fields didn't yet exist in the fields for this step so weren't ever
populated with a value.